rgw: ONLY move rgw_client_io_decoimpl.h to rgw_client_io_filters.h.
